refactor(terminal): 重构缓冲区模块结构
- 将 buffer.interface.ts 拆分为多个专用接口文件
- 创建独立的 buffer-cell.interface.ts 和 buffer-cell.ts 文件
- 提取 buffer-cursor.interface.ts 独立管理光标控制器
- 分离 buffer-view.interface.ts 和 buffer-view.ts 视图相关功能
- 创建 buffer-manager.interface.ts 和 buffer-manager.ts 管理器模块
- 添加统一的 buffer 模块入口文件 index.ts
- 更新导入路径以使用新的模块结构
- 移除旧的合并接口文件并优化代码组织
