--- frontend-workspace/terminal-ext/src/buffer.ts
@@ -1,428 +1,0 @@
-import { clamp } from "@xterm/xterm/src/vs/base/common/numbers.ts"
-import wcwidth from "wcwidth"
-import { AnsiCursor } from "./ansi-code.ts"
-import type {
-  IBufferCursorController,
-  IScreenBuffer,
-  IScreenBufferCell,
-  IScreenBufferManager,
-  IScreenBufferView,
-} from "./buffer.interface.ts"
-import type {
-  IDimension,
-  IRect,
-  IStyleProvider,
-  ITerminalStyle,
-  IWriteOption,
-} from "./capabilities.interface.ts"
-import { clipRect, TerminalStyle } from "./capabilities.ts"
-import type { ITextViewport } from "./text.interface.ts"
-import { type CursorPosition } from "./types.ts"
-import { assertValidCursorPosition, equal, isCursorPosition } from "./utils.ts"
-
-export class ScreenBufferCell implements IScreenBufferCell {
-  private _width = 0
-  private readonly _styles: ITerminalStyle[] = []
-  private _textStyle: ITerminalStyle = new TerminalStyle()
-  private readonly _style = new TerminalStyle()
-
-  constructor(
-    private readonly _row: number,
-    private readonly _col: number,
-    private _value: string = " ",
-  ) {
-    this._width = this._value[0] ? wcwidth(this._value[0]) : 0
-  }
-
-  copyFrom(other: ScreenBufferCell): this {
-    this._value = other.getValue()
-    this._width = other._width
-    this._textStyle.copyFrom(other._textStyle)
-    this._styles.length = 0
-    this._styles.push(...other._styles.map((i) => new TerminalStyle(i)))
-    return this
-  }
-
-  clone() {
-    const result = new ScreenBufferCell(this._row, this._col, this._value)
-    result._styles.push(...this._styles.map((i) => new TerminalStyle(i)))
-    result._textStyle.copyFrom(this._textStyle)
-    return result
-  }
-
-  appendStyle(style: ITerminalStyle): this {
-    this._styles.push(style)
-    return this
-  }
-
-  removeStyle(style: ITerminalStyle): this {
-    const index = this._styles.indexOf(style)
-    if (index >= 0) this._styles.splice(index, 1)
-    return this
-  }
-
-  setTextStyle(style: ITerminalStyle): this {
-    this._textStyle = style
-    return this
-  }
-
-  getTerminalStyle(): ITerminalStyle {
-    return this._style.reset().copyFrom(...this._styles, this._textStyle)
-  }
-
-  getWidth(): number {
-    return this._width
-  }
-
-  getValue(): string {
-    return this._value
-  }
-
-  setValue(val: string): this {
-    this._value = val[0] || " "
-    this._width = wcwidth(this._value) || 1
-    return this
-  }
-
-  getCol(): number {
-    return this._col
-  }
-
-  getRow(): number {
-    return this._row
-  }
-}
-
-export class ScreenBuffer implements IScreenBuffer {
-  private _matrix: Array<IScreenBufferCell[]> = []
-  private _oldStrings: string[][] = []
-  private readonly _backupMap = new Map<any, IScreenBufferCell[]>()
-  private _cursorOn: IScreenBufferCell
-  private _cursorPosUpdated = true
-
-  constructor(
-    private _rows: number,
-    private _cols: number,
-    readonly styleProvider: IStyleProvider,
-  ) {
-    this.resize({
-      getCols(): number {
-        return _cols
-      },
-      getRows(): number {
-        return _rows
-      },
-    })
-    this._cursorOn = this.getCell(0, 0)
-  }
-
-  getCursorController(): IBufferCursorController {
-    return null as any
-  }
-
-  cursorOn(pos: IScreenBufferCell | CursorPosition): this {
-    if (isCursorPosition(pos)) {
-      this._cursorOn = this.getCell(pos.row, pos.col)
-    } else {
-      this._cursorOn = pos
-    }
-    this._cursorPosUpdated = true
-    return this
-  }
-
-  getCursorOn(): IScreenBufferCell {
-    return this._cursorOn
-  }
-
-  save(key: any, range: IRect): this {
-    this._backupMap.set(
-      key,
-      this.getCells(range).map((i) => i.clone()),
-    )
-    return this
-  }
-
-  restore(key: any): this {
-    const cells = this._backupMap.get(key)
-    if (cells) {
-      for (let cell of cells) {
-        try {
-          this.getCell(cell.getRow(), cell.getCol()).copyFrom(cell)
-        } catch (e) {
-          console.error(e)
-        }
-      }
-      this._backupMap.delete(key)
-    }
-    return this
-  }
-
-  getCells(range?: IRect): IScreenBufferCell[] {
-    if (!range) return this._matrix.flat()
-    const array: IScreenBufferCell[] = []
-    const start = range.getStartPosition()
-    const end = range.getEndPosition()
-    for (let r = start.row; r <= end.row; r++) {
-      for (let c = start.col; c <= end.col; c++) {
-        array.push(this.getCell(r, c))
-      }
-    }
-    return array
-  }
-
-  getScreenBufferView(range?: IRect): IScreenBufferView {
-    const { _rows, _cols } = this
-    const r = range ?? {
-      getCols(): number {
-        return _cols
-      },
-      getRows(): number {
-        return _rows
-      },
-      getStartPosition(): Readonly<CursorPosition> {
-        return { row: 1, col: 1 }
-      },
-      getEndPosition(): Readonly<CursorPosition> {
-        return { row: _rows, col: _cols }
-      },
-    }
-    assertValidCursorPosition(r.getEndPosition(), r.getStartPosition())
-    return new ScreenBufferView(r, this)
-  }
-
-  markFullUpdate(): this {
-    this._oldStrings = []
-    for (let r = 0; r < this._rows; r++) {
-      const array: string[] = []
-      this._oldStrings.push(array)
-      for (let c = 0; c < this._cols; c++) {
-        array.push("")
-      }
-    }
-    return this
-  }
-
-  outputDiff(): string {
-    let array: string[] = []
-    const cursorOn = this._cursorPosUpdated
-      ? AnsiCursor.SET_POSITION(this._cursorOn.getRow(), this._cursorOn.getCol())
-      : ""
-    for (let r = 0; r < this._matrix.length; r++) {
-      const cells = this._matrix[r]
-      for (let c = 0; c < cells.length; ) {
-        const cell = cells[c]
-        const string = cell.getTerminalStyle().toString(cell.getValue(), true)
-        if (this._oldStrings[r][c] !== string) {
-          array.push(AnsiCursor.SET_POSITION(r + 1, c + 1) + string)
-          this._oldStrings[r][c] = string
-        }
-        const step = cell.getWidth() || 1
-        c += step
-      }
-    }
-    this._cursorPosUpdated = false
-    return array.join("") + cursorOn
-  }
-
-  resize(dimension: IDimension): this {
-    this._rows = dimension.getRows()
-    this._cols = dimension.getCols()
-    const oldCells = this._matrix.slice()
-    this._matrix = new Array(this._rows)
-    for (let row = 0; row < this._matrix.length; row++) {
-      if (row >= this._oldStrings.length) {
-        this._oldStrings.push([])
-      }
-      if (this._oldStrings[row].length > this._cols)
-        this._oldStrings[row] = this._oldStrings[row].slice(0, this._cols)
-      while (this._oldStrings[row].length < this._cols) {
-        this._oldStrings[row].push("")
-      }
-      const cells: IScreenBufferCell[] = []
-      this._matrix[row] = cells
-      if (row < oldCells.length) {
-        const old = oldCells[row]
-        cells.push(...old.slice(0, this._cols))
-      }
-      while (cells.length < this._cols) {
-        const cell = new ScreenBufferCell(row, cells.length)
-        cell.appendStyle(this.styleProvider.getTerminalStyle())
-        cells.push(cell)
-      }
-    }
-    return this
-  }
-
-  getCell(row: number, col: number): IScreenBufferCell {
-    if (row < 0 || col < 0) throw new Error("Invalid row or col")
-    const result = this._matrix[row]?.[col]
-    if (!result) throw new Error("Invalid row or col")
-    return result
-  }
-
-  getCols(): number {
-    return this._cols
-  }
-
-  getRows(): number {
-    return this._rows
-  }
-}
-
-export class ScreenBufferView implements IScreenBufferView {
-  constructor(
-    private _range: IRect,
-    private readonly _buffer: IScreenBuffer,
-    private readonly _style: ITerminalStyle = new TerminalStyle(),
-  ) {
-    this.setupCellStyle()
-  }
-
-  getCells(range?: IRect): IScreenBufferCell[] {
-    const rect = range ? clipRect(this._range, range) : this._range
-    return this._buffer.getCells(rect)
-  }
-
-  getCell(row: number, col: number): IScreenBufferCell {
-    row = clamp(row, 0, this._range.getRows() - 1)
-    col = clamp(col, 0, this._range.getCols() - 1)
-    const startPos = this._range.getStartPosition()
-    return this._buffer.getCell(row + startPos.row, col + startPos.col)
-  }
-
-  setRange(range: IRect): this {
-    if (equal(range, this.getRange())) return this
-    this.resetCellStyle()
-    this._range = range
-    return this.setupCellStyle()
-  }
-
-  setupCellStyle() {
-    for (let cell of this._buffer.getCells(this._range)) {
-      cell.appendStyle(this._style)
-    }
-    return this
-  }
-
-  resetCellStyle() {
-    for (let cell of this._buffer.getCells(this._range)) {
-      cell.removeStyle(this._style)
-    }
-    return this
-  }
-
-  dispose(): this {
-    return this.resetCellStyle()
-  }
-
-  getStyle(): ITerminalStyle {
-    return this._style
-  }
-
-  erase(): this {
-    for (let cell of this._buffer.getCells(this._range)) {
-      cell.setValue(" ")
-    }
-    return this
-  }
-
-  write(text: ITextViewport, option?: IWriteOption): this {
-    const region = text.getRegion()
-    for (
-      let r = 0;
-      r < region.endRow - region.startRow + 1 &&
-      r + this._range.getStartPosition().row <= this._range.getEndPosition().row;
-      r++
-    ) {
-      const chars = text.getChars(r + region.startRow)
-      const accWidth = chars.reduce((pre, cur) => pre + cur.width, 0)
-      let c = 0,
-        i = 0
-      if (option?.align === "center") {
-        const bias = Math.floor((this._range.getCols() - accWidth) / 2)
-        if (bias >= 0) c = bias
-      } else if (option?.align === "right") {
-        const bias = this._range.getCols() - accWidth
-        if (bias >= 0) {
-          c = bias
-        }
-      }
-      for (; i < chars.length; i++) {
-        const char = chars[i]
-        if (c + char.width <= this._range.getCols()) {
-          const cell = this._buffer.getCell(
-            r + this._range.getStartPosition().row,
-            c + this._range.getStartPosition().col,
-          )
-          cell.setValue(char.char).setTextStyle(char.style)
-          c += char.width || 1
-        } else {
-          break
-        }
-      }
-    }
-    return this
-  }
-
-  getScreenBufferView(range: IRect): IScreenBufferView {
-    assertValidCursorPosition(range.getStartPosition(), range.getEndPosition())
-    const startPosition: CursorPosition = {
-      row: range.getStartPosition().row + this._range.getStartPosition().row,
-      col: range.getStartPosition().col + this._range.getStartPosition().col,
-    }
-    const endPosition: CursorPosition = {
-      row: range.getEndPosition().row + this._range.getEndPosition().row,
-      col: range.getEndPosition().col + this._range.getEndPosition().col,
-    }
-    return new ScreenBufferView(
-      {
-        getEndPosition(): Readonly<CursorPosition> {
-          return endPosition
-        },
-        getRows(): number {
-          return range.getRows()
-        },
-        getStartPosition(): Readonly<CursorPosition> {
-          return startPosition
-        },
-        getCols(): number {
-          return range.getCols()
-        },
-      },
-      this._buffer,
-    )
-  }
-
-  getScreenBuffer(): IScreenBuffer {
-    return this._buffer
-  }
-
-  getRange(): Readonly<IRect> {
-    return this._range
-  }
-}
-
-export class ScreenBufferManager implements IScreenBufferManager {
-  private readonly _stack: IScreenBuffer[] = []
-
-  constructor(buffer: IScreenBuffer) {
-    this._stack.push(buffer)
-  }
-
-  getScreenBuffer(): IScreenBuffer {
-    return this._stack.at(-1)!
-  }
-
-  add(styleProvider: IStyleProvider): this {
-    const buffer = this.getScreenBuffer()
-    this._stack.push(new ScreenBuffer(buffer.getRows(), buffer.getCols(), styleProvider))
-    return this
-  }
-
-  pop() {
-    if (this._stack.length === 1) throw new Error("Cannot pop the last buffer")
-    this._stack.pop()
-    this.getScreenBuffer().markFullUpdate()
-    return this
-  }
-}
